requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
langchain==0.1.11
langchain-community
langchain_pinecone==0.0.3
langchain_text_splitters
flask
pypdf
python-dotenv
-e .